+++ src/components/pages/campaigns-create-approve/components/csv-upload/index.tsx
@@ -94,10 +94,7 @@ const CSVUploadPopup: FC<TProps> = ({
       </IconContainer>
         <PopupTitle>Upload .csv file with links</PopupTitle>
         <PopupText>
-          Read the information about the required file format <TextLink onClick={() => {
-            if (!tokenStandard) { return alert('No token standard chosen') }
-            downloadAssetsBlankCSV(tokenStandard, 'example')
-          }}>here</TextLink>
+          Read the information about the required file format <TextLink href='https://linkdrop-docs.notion.site/Using-CSV-to-add-tokens-217684e66db24b45b791a4a5a92c0585'>here</TextLink>
         </PopupText>
       </PopupFormContent>
       <Buttons>