+++ src/components/pages/common/aside/index.tsx
@@ -1,14 +1,24 @@
 import { TProps } from './types'
-import { FC } from 'react'
+import { FC, useState } from 'react'
 import {
   WidgetComponent,
   WidgetAside,
   ButtonsContainer,
-  WidgetTitle,
   WidgetSubtitle
 } from '../index'
+import Icons from 'icons'
 
-import { ButtonStyled } from './styled-components'
+import {
+  ButtonStyled, 
+  WidgetOptions,
+  WidgetTitle,
+  WidgetTitleFlex,
+  MiniPopupContainerStyled,
+  OptionsList,
+  OptionsListItem,
+  OptionsListBorder,
+  TitleLoader
+} from './styled-components'
 
 const Aside: FC<TProps> = ({
   children,
@@ -26,10 +36,58 @@ const Aside: FC<TProps> = ({
   } = {},
   title,
   subtitle,
+  options,
+  loading
 }) => {
+
+  const [ showOptions, setShowOptions ] = useState(false)
+
+  const defineHeaderContents = () => {
+    if (loading) {
+      return  <WidgetTitleFlex>
+        {title}
+        <TitleLoader />
+      </WidgetTitleFlex>
+    }
+    if (options) {
+      return <WidgetTitleFlex>
+        {title}
+        <WidgetOptions onClick={() => {
+          setShowOptions(!showOptions)
+        }}>
+          <Icons.OptionsIcon />
+          {showOptions && <MiniPopupContainerStyled onClose={() => { setShowOptions(false) }}>
+            <OptionsList>
+              {options.map(option => {
+                return <>
+                  <OptionsListItem
+                    disabled={option.disabled}
+                    onClick={() => {
+                      if (option.disabled) { return }
+                      setShowOptions(false)
+                      option.action && option.action()
+                    }}
+                  >
+                    {option.icon}{option.title}
+                  </OptionsListItem>
+                  {option.bordered && <OptionsListBorder />}
+                </>
+              })} 
+            </OptionsList>
+            
+          </MiniPopupContainerStyled>}
+        </WidgetOptions>
+      </WidgetTitleFlex>
+    }
+
+    return <WidgetTitle>
+      {title}
+    </WidgetTitle>
+  }
+
   return <WidgetAside>
     <WidgetComponent>
-      <WidgetTitle>{title}</WidgetTitle>
+      {defineHeaderContents()}
       {subtitle && <WidgetSubtitle>{subtitle}</WidgetSubtitle>}
       {children}
       {(backAction || nextAction) && <ButtonsContainer>