+++ src/helpers/define-proxy-contract-funds.tsx
@@ -1,0 +1,7 @@
+const defineProxyContractFunds = async (
+  contractAddress: string,
+  provider: any
+) => {
+  return await provider.getBalance(contractAddress)
+}
+export default defineProxyContractFunds