+++ src/helpers/index.js
@@ -48,9 +48,19 @@ import {
 } from './create-workers'
 import convertLinksContent from './convert-links-content'
 import downloadAssetsBlankCSV from './download-assets-blank-csv'
+import defineCampaignStatus from './define-campaign-status'
+import campaignPause from './campaign-pause'
+import campaignUnpause from './campaign-unpause'
+import campaignRefund from './campaign-refund'
+import defineProxyContractFunds from './define-proxy-contract-funds'
 
 export {
   shortenString,
+  defineProxyContractFunds,
+  campaignRefund,
+  campaignPause,
+  campaignUnpause,
+  defineCampaignStatus,
   downloadAssetsBlankCSV,
   convertLinksContent,
   checkERC721AssetsDataForMint,