+++ src/web-workers/qrs-worker.tsx
@@ -15,6 +15,7 @@ export class QRsWorker {
   ) {
     this.cb = cb
   }
+
   public prepareQRs (
     quantity: number,
     dashboard_key: string