+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Ledger Dashboard
 
+## 2.1.1
+- Options for campaign to pause, unpause and withdrawal. Also copy of campaign details to clipboard ability added
+- Added pending status to campaign while waiting for transaction complete
+- JSON Rpc urls moved to env variables
+- Added loader indicator while pause/unpause is pending
+- Added loader indicator while checking if campaign was approved before
+- Added withdraw ability check 
+
 ## 2.1.0